requirements.txt
python-telegram-bot==20.7
openai-whisper==20231117
moviepy==1.0.3
opencv-python==4.8.1.78
numpy==1.24.3
torch==2.1.0
torchaudio==2.1.0
librosa==0.10.1
scipy==1.11.4
Pillow==10.1.0
python-dotenv==1.0.0
